refactor(backend): move igdb utils to module
