+++ apps/backend/src/utils.rs
@@ -11,10 +11,7 @@ use scdb::Store;
 use sea_orm::{ActiveModelTrait, ActiveValue, ConnectionTrait, DatabaseConnection};
 use serde::de::{self, DeserializeOwned};
 use serde::{Deserialize, Serialize};
-use surf::{
-    http::headers::{AUTHORIZATION, USER_AGENT},
-    Client, Config, Url,
-};
+use surf::Client;
 use tokio::task::JoinSet;
 
 use crate::background::{
@@ -22,7 +19,6 @@ use crate::background::{
 };
 use crate::config::AppConfig;
 use crate::entities::user_to_metadata;
-use crate::graphql::AUTHOR;
 use crate::importer::ImporterService;
 use crate::media::resolver::MediaService;
 use crate::providers::{
@@ -189,78 +185,6 @@ pub fn get_now_timestamp() -> u128 {
         .duration_since(UNIX_EPOCH)
         .expect("Time went backwards")
         .as_millis()
-}
-
-pub mod igdb {
-    use std::{env, fs};
-
-    use serde_json::json;
-
-    use super::*;
-    use crate::{config::VideoGameConfig, graphql::PROJECT_NAME};
-
-    #[derive(Deserialize, Debug, Serialize)]
-    struct Credentials {
-        access_token: String,
-        expires_at: u128,
-    }
-
-    async fn get_access_token(config: &VideoGameConfig) -> Credentials {
-        let mut access_res = surf::post(&config.twitch.access_token_url)
-            .query(&json!({
-                "client_id": config.twitch.client_id.to_owned(),
-                "client_secret": config.twitch.client_secret.to_owned(),
-                "grant_type": "client_credentials".to_owned(),
-            }))
-            .unwrap()
-            .await
-            .unwrap();
-        #[derive(Deserialize, Serialize, Default, Debug)]
-        struct AccessResponse {
-            access_token: String,
-            token_type: String,
-            expires_in: u128,
-        }
-        let access = access_res
-            .body_json::<AccessResponse>()
-            .await
-            .unwrap_or_default();
-        let expires_at = get_now_timestamp() + (access.expires_in * 1000);
-        let access_token = format!("{} {}", access.token_type, access.access_token);
-        Credentials {
-            access_token,
-            expires_at,
-        }
-    }
-
-    // Ideally, I want this to use a mutex to store the client and expiry time.
-    // However for the time being we will read and write to a file.
-    pub async fn get_client(config: &VideoGameConfig) -> Client {
-        let path = env::temp_dir().join("igdb-credentials.json");
-        let access_token =
-            if let Some(mut credential_details) = read_file_to_json::<Credentials>(&path) {
-                if credential_details.expires_at < get_now_timestamp() {
-                    tracing::info!("Access token has expired, refreshing...");
-                    credential_details = get_access_token(config).await;
-                    fs::write(path, serde_json::to_string(&credential_details).unwrap()).ok();
-                }
-                credential_details.access_token
-            } else {
-                let creds = get_access_token(config).await;
-                fs::write(path, serde_json::to_string(&creds).unwrap()).ok();
-                creds.access_token
-            };
-        Config::new()
-            .add_header("Client-ID", config.twitch.client_id.to_owned())
-            .unwrap()
-            .add_header(USER_AGENT, format!("{}/{}", AUTHOR, PROJECT_NAME))
-            .unwrap()
-            .add_header(AUTHORIZATION, access_token)
-            .unwrap()
-            .set_base_url(Url::parse(&config.igdb.url).unwrap())
-            .try_into()
-            .unwrap()
-    }
 }
 
 pub mod openlibrary {
